feat: implement VideoShareModal component and integrate it into CallComponent and AboutComponent refactor: update cSpell configuration to include 'nodownload' in words list
